User profile interface and upload photo ability
