+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.biketrekteam.biketrekandroid" >
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
 
 
     <application
@@ -12,7 +13,7 @@
         android:theme="@style/AppTheme" >
         <meta-data android:name="com.facebook.sdk.ApplicationId" android:value="@string/facebook_app_id"/>
         <activity
-            android:name=".User_profile"
+            android:name=".UserProfileActivity"
             android:label="@string/title_activity_user_profile" >
         </activity>
         <activity
